src/requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2; python_full_version >= '3.5.3'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==19.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
boto3==1.13.21
botocore==1.16.21
chardet==3.0.4
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
idna==2.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
multidict==4.7.6; python_version >= '3.5'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
s3transfer==0.3.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
slackclient==2.7.0
urllib3==1.25.9
yarl==1.4.2; python_version >= '3.5'
